Move spec to inside lib/
